% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atomdat_2}
\alias{atomdat_2}
\title{Agent predictions for response 2}
\format{
A data frame with 616 rows and 5 columns.
\describe{
  \item{pmean}{Predictive mean.}
  \item{lpdens}{Log predictive density.}
  \item{method}{Aggregation method.}
  \item{t}{Time.}
  \item{ytrue}{True value of the response, not really implemented.}
}
}
\usage{
atomdat_2
}
\description{
Can be generated from oscbvar::macrodata, but is included here
  since that takes a lot of time. Agent predictions of GDPTCPI.
}
\keyword{datasets}
