% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atomdat_3}
\alias{atomdat_3}
\title{Agent predictions for response 3}
\format{
A data frame with 616 rows and 5 columns.
\describe{
  \item{pmean}{Predictive mean.}
  \item{lpdens}{Log predictive density.}
  \item{method}{Aggregation method.}
  \item{t}{Time.}
  \item{ytrue}{True value of the response, not really implemented.}
}
}
\usage{
atomdat_3
}
\description{
Bla bla bla.
}
\keyword{datasets}
