% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bikes_d_log}
\alias{bikes_d_log}
\title{Bikeshare data, daily}
\format{
A data frame with 730 rows and 13 columns
\describe{
 \item{logcnt}{Log of cnt.}
 \item{t}{Time.}
 \item{yr}{Year.}
 \item{mnth}{Month.}
 \item{workingday}{Weekday but not a holiday.}
 \item{temp}{Normalized temperature in celsius (how is this in 
     celsius...).}
 \item{hum}{Normalized humidity. I can't see how this is normalized.}
 \item{windspeed}{"Normalized" windspeed.}
 \item{sandy1}{Hurricane day 1.}
 \item{sandy2}{Hurricane day 2.}
 \item{weather_1}{1: clear, few clouds, partly cloudy. Baseline
     is light snow, light rain and thunderstorm and scattered clouds.}
 \item{weather_2}{2: mist and cloudy, mist and broken clouds.}
 \item{logcnt_l}{Lagged log count.}
}
}
\usage{
bikes_d_log
}
\description{
Data on bikeshare, daily version, with a bunch of covariates.
}
\keyword{datasets}
