% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes_nb_lm.R
\name{bikes_linreg}
\alias{bikes_linreg}
\title{Notebook generator: simple linear regression}
\usage{
bikes_linreg(log_scale = TRUE, agc = list(1, 200, TRUE))
}
\arguments{
\item{log_scale}{Should the dependent variable be on log scale?}

\item{agc}{List of atomic prediction generation controllers. The 
first element of the list gives the starting time (ie what  
observation is considered as t = 1), the second element is the 
minimum window length used for estimation, and the third one is a 
boolean indicating if the estimation window is rolling or not.}
}
\description{
Generates a notebook for a frequentist regression model.
}
\details{
lm() basically
}
