% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_nb_generators.R
\name{bikes_regression}
\alias{bikes_regression}
\title{Notebook generator: simple regression}
\usage{
bikes_regression(log_scale = FALSE, agc = list(1, 60, TRUE))
}
\arguments{
\item{log_scale}{Should the dependent variable be on log scale?}

\item{agc}{List of atomic prediction generation controllers. The 
first element of the list gives the starting time (ie what  
observation is considered as t = 1), the second element is the 
minimum window length used for estimation, and the third one is a 
boolean indicating if the estimation window is rolling or not.}
}
\description{
Generates a notebook for a bayesian regression model.
}
\details{
rstanarm default yes.
}
