% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikes_nb_svbvar.R
\name{bikes_svbvar}
\alias{bikes_svbvar}
\title{Notebook generator: Stochastic BVAR}
\usage{
bikes_svbvar(
  log_scale = TRUE,
  agc = list(1, 60, FALSE),
  include_intercept = TRUE
)
}
\arguments{
\item{log_scale}{Boolean. Do you want to take the log of the count
variable? Defaults to TRUE.}

\item{agc}{List of atomic prediction generation controllers. The 
first element of the list gives the starting time (ie what  
observation is considered as t = 1), the second element is the 
minimum window length used for estimation, and the third one is a 
boolean indicating if the estimation window is rolling or not.}

\item{include_intercept}{Should the design matrix include an
intercept? Defaults to TRUE.}
}
\description{
Generates a notebook for BVAR models with stochastic volatility
specifically for the bike sharing data. For some reason the svsample
function does not work when any of the columns in covariates have 
"to many zeroes" (more than half?), in which case you need to
specify "startpara = list(mu = 0, phi = 0, sigma = 1, beta = rep(0,
ncol(covariates)).".
}
\details{
Uses the stochvol package with default priors.
}
