% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_osa.R
\name{bvar_osa}
\alias{bvar_osa}
\title{Predictive density for BVAR models}
\usage{
bvar_osa(y, pd, logscale = TRUE)
}
\arguments{
\item{y}{The observed value for which we wish to calculate the predictive density.}

\item{pd}{A list obtained from bvar_pd with the predictive distribution.}

\item{logscale}{Whether to return the density on log scale or not. Defaults to TRUE.}
}
\description{
Multivariate predictive density for BVAR models. Takes an object from bvar_pd as input.
}
