% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_osa_marg.R
\name{bvar_osa_marg}
\alias{bvar_osa_marg}
\title{Marginal predictive density for BVAR models}
\usage{
bvar_osa_marg(y, pd, marg, logscale = TRUE)
}
\arguments{
\item{y}{The observed value for which we wish to calculate the predictive density.}

\item{pd}{A list obtained from bvar_pd with the predictive distribution.}

\item{marg}{Which variable to get the marginal for.}

\item{logscale}{Whether to return the density on log scale or not. Defaults to TRUE.}
}
\description{
Marginal predictive density for BVAR models. Takes an object from bvar_pd as input.
}
