% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvar_pd.R
\name{bvar_pd}
\alias{bvar_pd}
\title{Generates the predictive density for a BVAR}
\usage{
bvar_pd(Z, z, Y, p)
}
\arguments{
\item{Z}{Matrix of stacked covariates, with each row corresponding to one observation. Typically lagged values of Y.}

\item{z}{A vector of covariates for the observation we wish to predict.}

\item{Y}{A matrix of stacked dependent variables, with each row corresponding to one observation (a vector of (y_1t, ... , ymt)).}

\item{p}{Number of lags.}
}
\description{
Generates the one-step ahead predictive distribution for a BVAR model with Jeffreys' prior. Returns a list that can be used to get predicitive denistites usinng bvar_osa and bvar_osa_marg.
}
