% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliper_relevance.R
\name{caliper_relevance}
\alias{caliper_relevance}
\title{Relevance assessment method: Caliper}
\usage{
caliper_relevance(atomic_df, start_agg, tol, sotw)
}
\arguments{
\item{atomic_df}{Data frames with agent predictions.}

\item{start_agg}{From which value of t to start aggregating}

\item{tol}{Determines how similar observations need to be to be included.}

\item{sotw}{Data frame containing the state of the world at each time point.
The first column of this data frame should be t (as in time).}
}
\description{
Calculates weights for which previous log scores to use based on the 
relevance adjusted caliper method. All observations within a certain
distance are given equal weight, all outside of that distance are given
zero weight.
}
