% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dmdat}
\alias{dmdat}
\title{Additional decision maker data.}
\format{
A data frame with 293 rows and 2 variables:
\describe{
  \item{USSURV1055}{This is the one we use.}
  \item{USSURV1459}{This is the one we do not use.}
}
}
\usage{
dmdat
}
\description{
Bla bla bla.
}
\keyword{datasets}
