% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_gen_aggregate_preds.R
\name{gen_agg_preds}
\alias{gen_agg_preds}
\title{Generate aggregate predictions}
\usage{
gen_agg_preds(
  atomic_df,
  start_agg,
  sotw,
  baseline = TRUE,
  caliper = TRUE,
  mahala = TRUE,
  tol = 5,
  woc = "full"
)
}
\arguments{
\item{atomic_df}{Data frame containing the atomic predictions to
base the aggregate predictions on.}

\item{start_agg}{Which time point to generate the first aggregate
prediction for.}

\item{sotw}{Decision maker data set, observation one should
correspond to start_t used to generate the atomic predictions.}

\item{baseline}{Whether to generate the baseline aggregations
(gewisano and equal weights). Defaults to TRUE.}

\item{caliper}{Whether to generate RAP based on the caliper method.
Defaults to TRUE.}

\item{mahala}{Whether to generate RAP based on the mahalanobis
method. Defaults to TRUE.}

\item{tol}{Tolerance parameter for the caliper method. Defaults to
5.}

\item{woc}{Weight on caliper, parameter for the caliper method.
Defaults to "full."}
}
\description{
Super-function that generates aggregate predictions a la carte.
}
