% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_gen_all_data.R
\name{gen_all_data}
\alias{gen_all_data}
\title{Generate all the data needed for the macroeconomics example}
\usage{
gen_all_data(
  start_t = 5,
  start_agg = 161,
  rolling = FALSE,
  window_length = 60,
  bvar_3 = TRUE,
  bvar_7 = TRUE,
  bvar_3_o2 = TRUE,
  bvar_7_o2 = TRUE,
  svbvar_3 = TRUE,
  svbvar_7 = TRUE,
  svbvar_3_o2 = TRUE,
  svbvar_7_o2 = TRUE,
  bart_7 = TRUE,
  tvpsvbvar_3 = TRUE,
  bvar_3_basic = FALSE,
  bvar_7_basic = FALSE
)
}
\arguments{
\item{start_t}{Which observation in the macrodata to consider as t = 1.}

\item{start_agg}{At which time point to start aggregating. This will depend
on how much training data you want to feed the algorithms.}

\item{rolling}{Whether or not to use rolling estimation windows. Defaults to
FALSE.}

\item{window_length}{Minimum number of observations used for estimation.
Defaults to 60.}

\item{bvar_3}{Should the 3-dimensional VAR(1) model be included?}

\item{bvar_7}{Should the 7-dimensional VAR(1) model be included?}

\item{bvar_3_o2}{Should the 3-dimensional VAR(2) model be included?}

\item{bvar_7_o2}{Should the 7-dimensional VAR(2) model be included?}

\item{svbvar_3}{Should the 3-dimensional BVAR with stochastic volatility be
included?}

\item{svbvar_7}{Should the 7-dimensional BVAR with stochastic volatility be
included?}

\item{svbvar_3_o2}{Should the 3-dimensional VAR(2) with stochastic volatility
be included?}

\item{svbvar_7_o2}{Should the 7-dimensional VAR(2) with stochastic volatility
be included?}

\item{bart_7}{Should the 7-dimensional BART model be included?}

\item{tvpsvbvar_3}{Should the 3-dimensional time-varying-parameter BVAR with
stochastic volatility be included?}

\item{bvar_3_basic}{Should the basic 3-dimensional flat-Jeff VAR(1) model be
included?}

\item{bvar_7_basic}{Should the basic 7-dimensional flat-Jeff VAR(1) model be
included?}

\item{tol}{What tolerance to use in the caliper method.}

\item{woc}{What weight-on-caliper method to use in the caliper method.}
}
\description{
Generates all the data needed to create the figures and table
  in the macroeconomics example.
}
