% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers_general.R
\name{gen_atomic_list}
\alias{gen_atomic_list}
\title{Generate list of atomic models}
\usage{
gen_atomic_list()
}
\description{
Generates a list of all the available atomic models. Passed as
  an argument to the atomic prediction generator (and the gen_all_data fun).
}
