% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers_general.R
\name{gen_sotw}
\alias{gen_sotw}
\title{Generate state_of_the_world}
\usage{
gen_sotw(start_t)
}
\arguments{
\item{start_t}{Which observation to consider as t = 1.}
}
\description{
Generates a data frame with the state of the world, including the t-notation
used by the rest of the package.
}
