% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers.R
\name{gen_Z}
\alias{gen_Z}
\title{Generate Z matrix}
\usage{
gen_Z(data, start_t, lags, include_intercept)
}
\arguments{
\item{data}{Data frame (Y).}

\item{start_t}{Which observation is considered as t=1. Needs to be greater 
than the number of lags.}

\item{lags}{Order of the VAR.}

\item{include_intercept}{Boolean. Determines is an intercept should be used
or not.}
}
\description{
Generates a Z (design) matrix for a VAR model.
}
\details{
Given a data frame, a starting point, a number of lags, and whether to
include an intercept or not, creates a design matrix with correctly names
columns.
}
