% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_relevance_functions.R
\name{mahala_relevance}
\alias{mahala_relevance}
\title{Relevance assessment method: Mahalanobis}
\usage{
mahala_relevance(atomic_df, sotw, start_agg)
}
\arguments{
\item{atomic_df}{Data frames with agent predictions.}

\item{sotw}{Data frame containing the state of the world at each time point.
The first column of this data frame should be t (as in time).}

\item{start_agg}{From which value of t to start aggregating}
}
\description{
Calculates weights based on the Mahalanobis method.
}
