% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis_relevance.R
\name{mahalanobis_relevance}
\alias{mahalanobis_relevance}
\title{Mahalanobis relevance}
\usage{
mahalanobis_relevance(t, data)
}
\arguments{
\item{t}{Positive integer greater than 2. This is the timepoint (relative to the notebook) for which we wish to make our prediction.}

\item{data}{Notebook containing all predictions, outcomes, covariates, and predictive distributions. Should be generated by notebook_bvar.}
}
\description{
Calculates weights for which previous log scores to use based on the
Mahalanobis distance. All observations are included, but the weights
decrease quickly as a funcion of the Mahalanobis distance. Note that
there is no Mahalanobisification done, so you have to use a rescaled
data set!
}
