% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_bvar}
\alias{nb_bvar}
\title{Notebook generator: BVAR NiW}
\usage{
nb_bvar(
  data,
  window_length = 60,
  rolling = FALSE,
  start_t = 5,
  lags = 1,
  overall_tightness = 0.2,
  lag_decay = 1,
  include_intercept = TRUE
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook. Should include only
the variables used by the model, and the outcome variable of interest
should be in column 1.}

\item{window_length}{Minimum length of the estimation window. Defaults to
60 quarters (i.e. 5 years).}

\item{rolling}{Whether to use a rolling estimation window or not. Defaults to
FALSE.}

\item{start_t}{Which observation starts the estimation sample. Defaults to 5.}

\item{lags}{The order of the VAR model. Defaults to 1.}

\item{overall_tightness}{Overall tightness (pi_1 in Sunes notation). Defaults
to 0.2}

\item{lag_decay}{The lag decay rate. Defaults to 1.}

\item{include_intercept}{Whether or not to include an intercept. Defaults to
TRUE.}
}
\description{
Generates a notebook for a BVAR model with a Minnesota-flavoured NiW prior.
}
\details{
Generates a notebook of predictions for the decision maker to use. Uses a 
version of the Minnesota prior: the prior for the variance is data based
with nu_0 set to the number of time series plus two, and S_0 obtained by
running a simple AR(4) model and extracting the diagonal elements. The prior
for the regression coefficients is tweaked using two hyperparameters: the
overall tightness (defaults to 0.2) and the lag decay rate (deftaults to 1).
The cross-variable tightness is set to 1 to retain the Kronecker structure
required for conjugacy.
}
