% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_bvar_2.R
\name{nb_bvar_2}
\alias{nb_bvar_2}
\title{Generates a notebook for a BVAR model v2}
\usage{
nb_bvar_2(data, model, window_length = 60)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{model}{Which columns in the dataset should be included}

\item{window_length}{Length of estimation window. Defaults to 60.}
}
\description{
Uses Jeffreys' prior. Generates a notebook for the decision-maker to use. 
Uses a rolling window of default length 60 to estimate the model.
}
