% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_bvar_flat_Jeff}
\alias{nb_bvar_flat_Jeff}
\title{Notebook generator: simple (flat-Jeff) BVAR}
\usage{
nb_bvar_flat_Jeff(
  data,
  window_length = 60,
  rolling = FALSE,
  start_t = 5,
  lags = 1,
  include_intercept = FALSE
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{window_length}{Minimum length of the estimation window.}

\item{rolling}{Whether to use a rolling estimation window or not.}

\item{start_t}{Which observation to consider as t = 1. Defaults to 5.}

\item{lags}{The order of the VAR.}

\item{include_intercept}{Whether or not an intercept should be included.}
}
\description{
Generates a notebook for a BVAR model with flat-Jeff priors, meaning a flat
prior on the regression coefficients and Jeffreys' prior on the var-covariance.
}
