% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_svbvar}
\alias{nb_svbvar}
\title{Notebook generator: Stochastic BVAR}
\usage{
nb_svbvar(
  data,
  window_length = 60,
  rolling = FALSE,
  start_t = 5,
  lags = 1,
  include_intercept = TRUE
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook. Should include only
the variables used by the model, and the outcome varaible of interest
should be in column 1.}

\item{window_length}{Length of estimation window. Defaults to 60 quarters 
(i.e. 5 years).}

\item{rolling}{Whether to use a rolling estimation window or not. Defaults to
FALSE.}

\item{start_t}{Which observation to consider as t=1.}

\item{lags}{The order of the VAR model. Defaults to 1.}

\item{include_intercept}{Should the design matrix include an intercept? 
Defaults to TRUE.}
}
\description{
Generates a notebook for BVAR models with stochastic volatility.
}
\details{
Uses the stochvol package with default priors.
}
