% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nb_tvpbvar.R
\name{nb_tvpbvar}
\alias{nb_tvpbvar}
\title{Generates a notebook for a TVP-SV-BVAR model}
\usage{
nb_tvpbvar(
  data,
  model,
  window_length = 60,
  rolling = FALSE,
  nrep = 10000,
  nburn = 5000,
  tau = 20
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{model}{Which columns in the dataset should be included}

\item{window_length}{Minimum length of the estimation window.}

\item{rolling}{Whether to use a rolling estimation window or not.}

\item{nrep}{Number of MCMC draws (after burn-in)}

\item{nburn}{Number of burn-in draws}

\item{tau}{Number of observations to use for training prior.}
}
\description{
Uses default settings in bvarsv.
}
