% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_tvpsvbvar}
\alias{nb_tvpsvbvar}
\title{Notebook generator: TVP-SV-BVAR}
\usage{
nb_tvpsvbvar(data, agc, nrep = 10000, nburn = 5000, tau = 20)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{agc}{List of atomic prediction generation controllers. The first
element of the list gives the starting time (ie what observation is 
considered as t = 1), the second element is the minimum window length used
for estimation, and the third one is a boolean indicating if the estimation
window is rolling or not. Defaults to (5, 60, FALSE)}

\item{nrep}{Number of MCMC draws (after burn-in)}

\item{nburn}{Number of burn-in draws}

\item{tau}{Number of observations to use for training prior.}
}
\description{
Generates a notebook for a BVAR model with stochastic variance and time-
  varying parameters.
}
\details{
Uses default settings in bvarsv.
}
