% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_in_caliper.R
\name{no_in_caliper}
\alias{no_in_caliper}
\title{Count number of previous observation in caliper over time}
\usage{
no_in_caliper(cw, df, first_t, percentage)
}
\arguments{
\item{cw}{Caliper width value.}

\item{df}{Data frame containing the pooling variables. The first
column needs to be called "t" and contain the timepoint.}

\item{first_t}{Timepoint ("t") from which to start counting.}

\item{percentage}{Should the number of observations within the 
caliper (FALSE) or the percentage (TRUE) be returned?}
}
\value{
A data.frame with columns count (numer of observation within
  caliper), cw_val (caliper width), and time (time point).
}
\description{
Description goes here.
}
\details{
Details go here.
}
