% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook_bvar.R
\name{notebook_bvar}
\alias{notebook_bvar}
\title{Generates a notebook for a BVAR model}
\usage{
notebook_bvar(data, model, window_length = 60)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{model}{Which columns in the dataset should be included}

\item{window_length}{Length of estimation window. Defaults to 60.}
}
\description{
Uses Jeffreys' prior. Generates a notebook for the decision-maker to use. 
Uses a rolling window of default length 60 to estimate the model.
}
