% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers.R
\name{omega_minnesota}
\alias{omega_minnesota}
\title{Omega_0 for the Minnesota prior}
\usage{
omega_minnesota(lags, overall_tightness, lag_decay, m, include_intercept)
}
\arguments{
\item{overall_tightness}{Overall tightness.}

\item{lag_decay}{Lag decay.}

\item{m}{Dimensionality of each y-vector (or columns in Y).}

\item{include_intercept}{Whether or not the model includes an intercept.}
}
\description{
Generates an Omega_0 according to a (the) Minnesota prior.
}
\details{
Prior variance for the intercept (if included) is set to 10 000. Cross series
tightness is set to 1.
}
\keyword{internal}
