\name{oscbvar-package}
\alias{oscbvar-package}
\alias{oscbvar}
\docType{package}
\title{
\packageTitle{oscbvar}
}
\description{
\packageDescription{oscbvar}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{oscbvar}
\packageIndices{oscbvar}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{oscbvar}

Maintainer: \packageMaintainer{oscbvar}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
