% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_in_caliper.R
\name{plot_no_in_caliper}
\alias{plot_no_in_caliper}
\title{Plot number of previous observation in caliper over time}
\usage{
plot_no_in_caliper(cw_list, df, first_t, percentage = TRUE)
}
\arguments{
\item{cw_list}{List or vector of caliper width value.}

\item{df}{Data frame containing the pooling variables. The first
column needs to be called "t" and contain the timepoint.}

\item{first_t}{Timepoint ("t") from which to start counting.}

\item{percentage}{Should the number of observations within the 
caliper (FALSE) or the percentage (TRUE) be returned?}
}
\value{
A ggplot object.
}
\description{
Description goes here.
}
\details{
Details go here.
}
