% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers_aggregation.R
\name{RAL_calculator}
\alias{RAL_calculator}
\title{Local logscore calculator}
\usage{
RAL_calculator(weight_df, atomic_df)
}
\arguments{
\item{weight_df}{Data frame containing relevance adjusted weights.
Generated by a relevance function (caliper relevance or
mahalanobis atm).}

\item{atomic_df}{Data frame with atomic predictions from any number
of models.}
}
\description{
Calculates the local logscores given a weight data frame and a data
frame of atomic predictions. It does not calculate an estimate of
ELPD, rather N*ELPD, where N is the bumber of observations, so that 
the weights scale with weight of evidence.
}
