% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_goldfish.R
\name{relevance_goldfish}
\alias{relevance_goldfish}
\title{The relevance adjusted goldfish}
\usage{
relevance_goldfish(t, notebook, tol)
}
\arguments{
\item{t}{Decides which observation represents the present (in the notebook)}

\item{notebook}{Notebook containing all predictions}

\item{tol}{Determines how similar observations need to be to be included}
}
\description{
Calculates weights for which previous log scores to use based on the 
relevance adjusted goldfish.
}
