% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_goldfish_2.R
\name{relevance_goldfish_2}
\alias{relevance_goldfish_2}
\title{The relevance adjusted goldfish v2}
\usage{
relevance_goldfish_2(t, data, tol)
}
\arguments{
\item{t}{Positive integer greater than 2. This is the timepoint (relative to the notebook) for which we wish to make our prediction.}

\item{data}{Notebook containing all predictions, outcomes, covariates, and predictive distributions. Should be generated by notebook_bvar.}

\item{tol}{Determines how similar observations need to be to be included.}
}
\description{
Calculates weights for which previous log scores to use based on the 
relevance adjusted goldfish. The foldfish forgets everything that happened
more than ten minutes ago (in covariate space).
}
