% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers_aggregation.R
\name{selbest_weighting}
\alias{selbest_weighting}
\title{Weighting by picking the best model}
\usage{
selbest_weighting(data, sim_measure)
}
\arguments{
\item{data}{Which dataset to use.}

\item{sim_measure}{Caliper or mahala.}
}
\description{
Weighting by picking the best model
}
