% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tvpsvbvar_resp3}
\alias{tvpsvbvar_resp3}
\title{Predictions from the tvpsvbvar model with variable 3 as response.}
\format{
A data frame with 154 rows and 5 variables:
\describe{
  \item{pmean}{Predictive mean.}
  \item{lpdens}{Log predictive density.}
  \item{method}{Name of the model used.}
  \item{t}{Time point}
  \item{ytrue}{True value of the response value.}
}
}
\usage{
tvpsvbvar_resp3
}
\description{
Bla bla bla.
}
\keyword{datasets}
