% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_nb_generators.R
\name{bikes_bart}
\alias{bikes_bart}
\title{Notebook generator: BART}
\usage{
bikes_bart(
  covariates,
  log_scale = FALSE,
  agc = list(1, 60, TRUE),
  include_intercept = FALSE,
  nrep = 10000,
  nburn = 5000
)
}
\arguments{
\item{covariates}{Covarites, aka design matrix.}

\item{log_scale}{Should the dependent variable be on log scale?}

\item{agc}{List of atomic prediction generation controllers. The 
first element of the list gives the starting time (ie what  
observation is considered as t = 1), the second element is the 
minimum window length used for estimation, and the third one is a 
boolean indicating if the estimation window is rolling or not.}

\item{include_intercept}{Whether to include an intercep in the
design matrix or not. Defaults to FALSE since the intercept breaks
the function.}

\item{nrep}{Number of MCMC draws (after burn-in)}

\item{nburn}{Number of burn-in draws}
}
\description{
Generates a notebook for a BART model using the bike sharing data.
}
\details{
Uses default settings in dbarts.
}
