% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bike_nb_generators.R
\name{bikes_svbvar}
\alias{bikes_svbvar}
\title{Notebook generator: Stochastic BVAR}
\usage{
bikes_svbvar(
  covariates,
  log_scale = FALSE,
  agc = list(1, 60, FALSE),
  include_intercept = TRUE
)
}
\arguments{
\item{covariates}{Covariates to use in the model, in additional to
the count variable (which is always included). No defaults.}

\item{log_scale}{Boolean. Do you want to take the log of the count
variable? Defaults to FALSE.}

\item{agc}{List of atomic prediction generation controllers. The 
first element of the list gives the starting time (ie what  
observation is considered as t = 1), the second element is the 
minimum window length used for estimation, and the third one is a 
boolean indicating if the estimation window is rolling or not.}

\item{include_intercept}{Should the design matrix include an
intercept? Defaults to TRUE.}
}
\description{
Generates a notebook for BVAR models with stochastic volatility
specifically for the bike sharing data.
}
\details{
Uses the stochvol package with default priors.
}
