% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_relevance_functions.R
\name{caliper_relevance}
\alias{caliper_relevance}
\title{Relevance assessment method: Caliper}
\usage{
caliper_relevance(atomic_df, sotw, start_agg = 161, tol = 5, woc = "full")
}
\arguments{
\item{atomic_df}{Data frames with agent predictions.}

\item{sotw}{Data frame containing the state of the world at each time point.
The first column of this data frame should be t (as in time).}

\item{start_agg}{From which value of t to start aggregating}

\item{tol}{Determines how similar observations need to be to be included.}

\item{woc}{Weight on caliper. This specifies how the observations "within"
the caliper should be weighted. Takes vaules "full" meaning that only
observations inside the tolerance count, or "progressive", which gives
weight to the observations within the caliper proportional to the number of
observations within the caliper. Not that if the "full" method is being
used and error will be thrown if some timepoints have zero previous
observations within the caliper. "uncond" reverts to the unconditional
predictive ability when there is no data (ie giving each previous
observation equal weight).}
}
\description{
Calculates weights for which previous log scores to use based on the
relevance adjusted caliper method. All observations within a certain
distance are given equal weight, all outside of that distance are given
zero weight.
}
\details{
A potential problem with the caliper method is that we don't really know
how many observations are going to end up within the tolerance. I have no
idea what is going to turn out as a "reasonable" number, but clearly having
very few (1-5 say) observations would make the variance of the RAL very high.
}
