% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliper_relevance.R
\name{caliper_relevance}
\alias{caliper_relevance}
\title{The relevance adjusted goldfish v2}
\usage{
caliper_relevance(t, data, tol)
}
\arguments{
\item{t}{Positive integer greater than 2. This is the timepoint (relative to the notebook) for which we wish to make our prediction.}

\item{data}{Notebook containing all predictions, outcomes, covariates, and predictive distributions. Should be generated by notebook_bvar.}

\item{tol}{Determines how similar observations need to be to be included.}
}
\description{
Calculates weights for which previous log scores to use based on the 
relevance adjusted caliper method. All observations within a certain
distance are given equal weight, all outside of that distance are given
zero weight.
}
