% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliper_relevance.R
\name{caliper_relevance}
\alias{caliper_relevance}
\title{Caliper method for local weights}
\usage{
caliper_relevance(atomic_df, sotw, start_agg = 161, cw = 5, mvc = 1)
}
\arguments{
\item{atomic_df}{Data frames with agent predictions.}

\item{sotw}{Data frame containing the state of the world at each
time point, which can include decision maker variables not in any
of the atomic models. The first column of this data frame should
be t (as in time).}

\item{start_agg}{From which value of t to start aggregating, ie 
producing aggregate predictions.}

\item{cw}{The caliper width.}

\item{mvc}{Minimum viable cluster size.}
}
\value{
A data table that consists of several stacked data tables
  (each indexed by the column t). Each subtable has a value t2 for
  each previous observation, and to each of those a corresponding
  similarity.
}
\description{
Calculates weights for which previous log scores to use based on the
caliper method.
}
\details{
The caliper method splits the estimate between a local and a global
part. The local part is the sum of all log scores within the 
caliper width, while the global part is the global average. The
balance between the global and local part depends on the minimum
viable cluster size. If the number of observations within the
caliper equals or exceeds the minimum viable cluster size, the
global estimates gets zero weight. When there are no obsevations 
within the cluster, the global estimate gets all the weight. For all
situations between these extremes, a linear combination depending on
how large a percentage of the minimum viable cluster size is
attained. (Se paper for maths.)
}
