% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_gen_all_data.R
\name{gen_all_data}
\alias{gen_all_data}
\title{Generate all the data needed for the macroeconomics example}
\usage{
gen_all_data(
  model_list,
  start_t = 5,
  start_agg = 161,
  rolling = FALSE,
  window_length = 60,
  tol = 5,
  woc = "full"
)
}
\arguments{
\item{model_list}{List of names of the atomic models to use.}

\item{start_t}{Which observation in the macrodata to consider as t = 1.}

\item{start_agg}{At which time point to start aggregating. This will depend
on how much training data you want to feed the algorithms.}

\item{rolling}{Whether or not to use rolling estimation windows. Defaults to
FALSE.}

\item{window_length}{Minimum number of observations used for estimation.
Defaults to 60.}

\item{tol}{Parameter that determines caliper tolerance.}

\item{woc}{Parameter that determines "nonexistence" problems for the caliper.}
}
\description{
Generates all the data needed to create the figures and table
  in the macroeconomics example.
}
