% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks.R
\name{gen_atomic_df}
\alias{gen_atomic_df}
\title{Generates an empty data frame to store atomic predictions}
\usage{
gen_atomic_df()
}
\description{
Generates an empty data frame to store atomic predictions
}
\keyword{internal}
