% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers.R
\name{gen_atomic_df}
\alias{gen_atomic_df}
\title{Generate atomic data frame}
\usage{
gen_atomic_df()
}
\description{
Generates an empty data frame to store atomic predictions.
}
\details{
Separate from the actual notebook functions to ensure that each function
stores its predictions in identical data frames, since otherwise the merging
won't work.
}
\keyword{internal}
