% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_gen_atomic_preds.R
\name{gen_atomic_preds}
\alias{gen_atomic_preds}
\title{Generate atomic predictions}
\usage{
gen_atomic_preds(model_list, window_length = 60, rolling = FALSE, start_t = 5)
}
\arguments{
\item{model_list}{List of names of the atomic models to use.}

\item{window_length}{Minimum number of observations used for estimation.
Defaults to 60.}

\item{rolling}{Whether to use a rolling window for estimation, defaults to
false.}

\item{start_t}{Which observation to consider the first real observation. All
observations prior to that one will be considered fixed. Set to 5 by
default to allow the specification of models that are autoregressive of
order 4 without breaking anything.}
}
\description{
Generates a notebook of atomic predictive densities and means.
}
\details{
Function to generate the atomic predictions (that is the predictions of the
individual models, not of aggregation scehemes) for the macro-data example 
for predicting DGP. Essentialy a list of potential models to include that the
user selects from. Also contains some global settings that always should be
the same between models, such as estimation window length, starting time, and
which data set to use. Uses the data set previously known as og_medium_scaled
so to minimize the risk of accidentally using the wrong data set...
}
