% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks.R
\name{gen_atomic_preds}
\alias{gen_atomic_preds}
\title{Generates a notebook of atomic predictive densities and means}
\usage{
gen_atomic_preds(
  data,
  window_length = 60,
  rolling = FALSE,
  bvar_3 = TRUE,
  bvar_7 = TRUE,
  svbvar_3 = TRUE,
  svbvar_7 = TRUE,
  bart_7 = TRUE,
  tvpsvbvar_3 = TRUE
)
}
\arguments{
\item{data}{Specify dataset, should be based on OG medium, this is just for 
the macro data example.}

\item{window_length}{How many observations to use for estimation}

\item{rolling}{Whether to use a rolling window for estimation, defaults to false}

\item{bvar_3}{Should the 3-dimensional BVAR model be included?}

\item{bvar_7}{Should the 7-dimensional BVAR model be included?}

\item{svbvar_3}{Should the 3-dimensional BVAR with stochastic volatility be included?}

\item{svbvar_7}{Should the 7-dimensional BVAR with stochastic volatility be included?}

\item{bart_7}{Should the 7-dimensional BART model be included?}

\item{tvpsvbvar_3}{Should the 3-dimensional time-varying-parameter BVAR with stochastic
volatility be included?}
}
\description{
Function to generate the atomic predictions (that is the predictions of the
individual models, not of aggregation scehemes) for the macro-data example 
for predicting DGP.
}
