% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_gen_atomic_preds.R
\name{gen_atomic_preds}
\alias{gen_atomic_preds}
\title{Generate atomic predictions}
\usage{
gen_atomic_preds(model_list, agc = list(5, 60, FALSE, 1))
}
\arguments{
\item{model_list}{List of names of the atomic models to use.}

\item{agc}{List of atomic prediction generation controllers. The first
element of the list gives the starting time (ie what observation is 
considered as t = 1), the second element is the minimum window length used
for estimation, and the third one is a boolean indicating if the estimation
window is rolling or not.}
}
\description{
Generates a notebook of atomic predictive densities and means.
}
\details{
Function to generate the atomic predictions (that is the predictions of the
individual models, not of aggregation schemes) for the macro-data example 
for predicting DGP. Essentialy a list of potential models to include that the
user selects from. Also contains some global settings that always should be
the same between models, such as estimation window length, starting time, and
which data set to use. Uses the data set previously known as og_medium_scaled
so to minimize the risk of accidentally using the wrong data set...
}
