% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks.R
\name{nb_bart}
\alias{nb_bart}
\title{Generates a notebook for a BART model}
\usage{
nb_bart(
  data,
  model,
  window_length = 60,
  rolling = FALSE,
  nrep = 10000,
  nburn = 5000
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{model}{Which columns in the dataset should be included}

\item{window_length}{Minimum length of the estimation window.}

\item{rolling}{Whether to use a rolling estimation window or not.}

\item{nrep}{Number of MCMC draws (after burn-in)}

\item{nburn}{Number of burn-in draws}
}
\description{
Uses default settings in dbarts.
}
