% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_bart}
\alias{nb_bart}
\title{Notebook generator: BART}
\usage{
nb_bart(
  data = macrodata[, 1:7],
  agc = list(5, 60, TRUE, 1),
  lags = 1,
  include_intercept = FALSE,
  nrep = 10000,
  nburn = 5000
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook. Defaults to
the macroeconomic data set from this package using all variables.}

\item{agc}{List of atomic prediction generation controllers. The 
first element of the list gives the starting time (ie what  
observation is considered as t = 1), the second element is the 
minimum window length used for estimation, and the third one is a 
boolean indicating if the estimation window is rolling or not, the 
forth indicates which variable is the response variable and it 
defaults to 1 (GDP).}

\item{lags}{The order of the VAR. Defaults to 1.}

\item{include_intercept}{Whether to include an intercep in the
design matrix or not. Defaults to FALSE since the intercept breaks
the function.}

\item{nrep}{Number of MCMC draws (after burn-in)}

\item{nburn}{Number of burn-in draws}
}
\description{
Generates a notebook for a BART model.
}
\details{
Uses default settings in dbarts.
}
