% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_bvar_flat_Jeff}
\alias{nb_bvar_flat_Jeff}
\title{Generates a notebook for a BVAR simple version}
\usage{
nb_bvar_flat_Jeff(data, window_length = 60, rolling = FALSE)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{window_length}{Minimum length of the estimation window.}

\item{rolling}{Whether to use a rolling estimation window or not.}
}
\description{
Uses a flat Jeffreys' prior. Generates a notebook for the decision-maker to use. No intercept. One lag. Very simple.
}
