% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebooks.R
\name{nb_svbvar}
\alias{nb_svbvar}
\title{Generates a notebook for a simple stochastic volatility model}
\usage{
nb_svbvar(data, model, window_length = 60, rolling = FALSE)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{model}{Which columns in the dataset should be included as covariates.}

\item{window_length}{Length of estimation window. Defaults to 60.}

\item{rolling}{Whether to use a rolling estimation window or not.}
}
\description{
Uses Gregors prior. Generates a notebook for the decision-maker to use. 
Uses a rolling window of default length 60 to estimate the model.
}
