% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_nb_generators.R
\name{nb_tvpsvbvar}
\alias{nb_tvpsvbvar}
\title{Generates a notebook for a TVP-SV-BVAR model}
\usage{
nb_tvpsvbvar(
  data,
  model,
  window_length = 60,
  rolling = FALSE,
  start_t = 5,
  nrep = 10000,
  nburn = 5000,
  tau = 20
)
}
\arguments{
\item{data}{Dataset from which to generate the notebook}

\item{model}{Which columns in the dataset should be included}

\item{window_length}{Minimum length of the estimation window.}

\item{rolling}{Whether to use a rolling estimation window or not.}

\item{start_t}{Which observation to consider as t=1.}

\item{nrep}{Number of MCMC draws (after burn-in)}

\item{nburn}{Number of burn-in draws}

\item{tau}{Number of observations to use for training prior.}
}
\description{
Uses default settings in bvarsv.
}
