% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers_aggregation.R
\name{propto_weighting}
\alias{propto_weighting}
\title{Weighting proportional to RAL}
\usage{
propto_weighting(data)
}
\arguments{
\item{data}{Dataset to use}
}
\description{
Weighting proportional to RAL
}
