% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relevance_goldfish.R
\name{relevance_goldfish}
\alias{relevance_goldfish}
\title{The relevance adjusted goldfish}
\usage{
relevance_goldfish(t, notebook, tol)
}
\arguments{
\item{t}{Positive integer greater than 2. This is the timepoint (relative to the notebook) for which we wish to make our prediction.}

\item{notebook}{Notebook containing all predictions, outcomes, covariates, and predictive distributions. Should be generated by notebook_bvar.}

\item{tol}{Determines how similar observations need to be to be included.}
}
\description{
Calculates weights for which previous log scores to use based on the 
relevance adjusted goldfish.
}
