% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_helpers_aggregation.R
\name{selbest_weighting}
\alias{selbest_weighting}
\title{Weighting by picking the best model}
\usage{
selbest_weighting(data)
}
\arguments{
\item{data}{Which dataset to use}
}
\description{
Weighting by picking the best model
}
